git-svn-id: https://svn.neo4j.org/components/neo/trunk@2846 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ advanced/src/java/org/neo4j/impl/nioneo/store/Store.java
@@ -36,7 +36,7 @@ public interface Store
      */
     public int nextId();
     
-    public int getHighestPossibleIdInUse();
+    public long getHighestPossibleIdInUse();
 
-    public int getNumberOfIdsInUse();
+    public long getNumberOfIdsInUse();
 }
